simplified tests, now that monkeypatching not needed
