+++ Gemfile
@@ -1,3 +1,6 @@
 source 'https://rubygems.org'
 
+# move this into jekyll-pandoc.gemspect when it is released
+gem 'jekyll', :github => "jekyll/jekyll", :tag => "v3.0.0.pre.beta8"
+
 gemspec
