+++ jekyll-pandoc.gemspec
@@ -1,9 +1,8 @@
 # -*- encoding: utf-8 -*-
-require File.expand_path("../lib/jekyll-pandoc/version", __FILE__)
 
 Gem::Specification.new do |s|
   s.name        = 'jekyll-pandoc'
-  s.version     = JekyllPandoc::VERSION
+  s.version     = '0.1.0'
   s.platform    = Gem::Platform::RUBY
   s.authors     = ["Martin Fenner"]
   s.email       = 'mf@martinfenner.org'
@@ -14,11 +13,12 @@ Gem::Specification.new do |s|
 
   s.required_rubygems_version = ">= 1.3.6"
 
-  s.add_dependency "jekyll", ">=  1.4.3"
+  # TODO: uncomment when it is available 
+  # s.add_dependency "jekyll", '>= 3.0.0'
   s.add_dependency "pandoc-ruby", "~> 0.7.5"
   s.add_development_dependency 'rake'
-  s.add_development_dependency "rspec", "~> 2.6"
-  s.add_development_dependency "rdiscount", "~> 2.1.7"
+  s.add_development_dependency "rspec", "> 2.6"
+  s.add_development_dependency "rdiscount", "> 2.1.7"
 
   s.files       = Dir.glob("lib/**/*.rb")
 end
