--- lib/jekyll-pandoc/version.rb
@@ -1,3 +1,0 @@
-module JekyllPandoc
-  VERSION = "0.0.8"
-end
