+++ spec/citations.bib
@@ -1,0 +1,9 @@
+@article{Bush:1945,
+        Author = {Bush, Vannevar},
+        Date-Added = {2015-08-31 01:17:13 +0000},
+        Date-Modified = {2015-09-05 09:59:13 +0000},
+        Journal = {Atlantic},
+        Month = {July},
+        Title = {As We May Think},
+        Url = {http://www.theatlantic.com/magazine/archive/1945/07/as-we-may-think/303881/},
+        Year = {1945}}
